+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Removed installation of M365DSC, has to be taken care of outside of this module
+- Added check if the required Microsoft365DSC version is already installed, before trying to install it.
 
 ## [0.1.3] - 2023-11-13